Switch database from sqlite to postgre
There was trouble migrating some of the model changes, I'm hoping that
changing to PostgreSQL will help on that front. In addition, using
PostgreSQL seems to be the standard for django apps, so in order to keep
with tradition, I'll go ahead and do the same until I can think of a
good reason to switch.

Also migrated all data from sqlite database to the new postgre database.