Add an html_body field to BasicPost
Because the html_body() function had to be called whenever the post
needed to be displayed, this would use up a lot of processing power. As
a solution, an additional field containing the preprocessed html was
added to BasicPost to save on processing being done.

Before switching over to PostgreSQL, this change was being very
difficult, after switching over this was done easily.